refactor: 온보딩 검증 로직 분리 및 스타일 관련 경고 제거 (#12)
